feat:  unversioned version
- Added redoc
- Removed routeprefix to empty
