+++ src/Extensions/SwaggerExtensions.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.PlatformAbstractions;
 using Microsoft.OpenApi.Models;
 using PlusUltra.Swagger.Filters;
 using Swashbuckle.AspNetCore.Filters;
+using Swashbuckle.AspNetCore.ReDoc;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using Swashbuckle.AspNetCore.SwaggerUI;
 
@@ -14,7 +15,8 @@ namespace PlusUltra.Swagger.Extensions
 {
     public static class SwaggerExtensions
     {
-        public static IServiceCollection AddDocumentation(this IServiceCollection services, OpenApiInfo info, Action<SwaggerGenOptions> configuration = null)
+
+        public static IServiceCollection AddVersionedDocumentation(this IServiceCollection services, OpenApiInfo info, Action<SwaggerGenOptions> configuration = null)
         {
             services.AddSwaggerGen(
                 options =>
@@ -39,14 +41,37 @@ namespace PlusUltra.Swagger.Extensions
                     options.OperationFilter<AuthResponsesOperationFilter>();
                     // add a custom operation filter which sets default values
                     options.OperationFilter<SwaggerDefaultValues>();
-                    
+
                     configuration?.Invoke(options);
                 });
 
             return services;
         }
 
-        public static IApplicationBuilder UseDocumentation(this IApplicationBuilder app, IApiVersionDescriptionProvider provider, Action<SwaggerUIOptions> configuration = null)
+        public static IServiceCollection AddDocumentation(this IServiceCollection services, OpenApiInfo info, string groupName = "v1", Action<SwaggerGenOptions> configuration = null)
+        {
+            services.AddSwaggerGen(
+                options =>
+                {
+                    options.DescribeAllParametersInCamelCase();
+
+                    options.SwaggerDoc(groupName, info);
+
+                    // integrate xml comments
+                    IncludeXMLS(options);
+
+                    options.OperationFilter<AddResponseHeadersFilter>();
+                    options.OperationFilter<AuthResponsesOperationFilter>();
+                    // add a custom operation filter which sets default values
+                    options.OperationFilter<SwaggerDefaultValues>();
+
+                    configuration?.Invoke(options);
+                });
+
+            return services;
+        }
+
+        public static IApplicationBuilder UseVersionedDocumentation(this IApplicationBuilder app, IApiVersionDescriptionProvider provider, Action<SwaggerUIOptions> configuration = null)
         {
             // Enable middleware to serve generated Swagger as a JSON endpoint.
             app.UseSwagger();
@@ -58,9 +83,34 @@ namespace PlusUltra.Swagger.Extensions
                 // build a swagger endpoint for each discovered API version
                 foreach (var description in provider.ApiVersionDescriptions)
                 {
-                    c.SwaggerEndpoint($"./swagger/{description.GroupName}/swagger.json", description.GroupName.ToUpperInvariant());
+                    c.SwaggerEndpoint($"/swagger/{description.GroupName}/swagger.json", description.GroupName.ToUpperInvariant());
                     c.RoutePrefix = string.Empty;
                 }
+
+                configuration?.Invoke(c);
+            });
+
+            return app;
+        }
+
+        public static IApplicationBuilder UseDocumentation(this IApplicationBuilder app, string groupName = "v1", Action<ReDocOptions> configuration = null)
+        {
+            // Enable middleware to serve generated Swagger as a JSON endpoint.
+            app.UseSwagger();
+
+            app.UseReDoc(c =>
+            {
+                c.SpecUrl($"/swagger/{groupName}/swagger.json");
+                c.RoutePrefix = "docs";
+                c.HideHostname();
+                c.HideDownloadButton();
+                c.ExpandResponses("200,201");
+                c.RequiredPropsFirst();
+                c.NoAutoAuth();
+                c.PathInMiddlePanel();
+                c.NativeScrollbars();
+                c.OnlyRequiredInSamples();
+                c.SortPropsAlphabetically();
 
                 configuration?.Invoke(c);
             });
